ocr-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
pillow>=9.0.0
opencv-python-headless>=4.5.0
numpy>=1.20.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.1+cpu
torchvision==0.22.1+cpu
easyocr>=1.6.0